Convert new examples in libraries/validation.
