+++ user_guide_src/source/libraries/validation/028.php
@@ -1,3 +1,19 @@
 <?php
 
-$error = $validation->getError('username');
+// for data
+'contacts' => [
+    'friends' => [
+        [
+            'name' => 'Fred Flinstone',
+        ],
+        [
+            'name' => '',
+        ],
+    ]
+];
+
+// rule
+// contacts.*.name => 'required'
+
+// error will be
+//'contacts.friends.1.name' => 'The contacts.*.name field is required.',
