+++ user_guide_src/source/libraries/validation/029.php
@@ -1,5 +1,3 @@
 <?php
 
-if ($validation->hasError('username')) {
-    echo $validation->getError('username');
-}
+$error = $validation->getError('username');
