+++ user_guide_src/source/libraries/validation/030.php
@@ -1,7 +1,5 @@
-<div class="alert alert-danger" role="alert">
-    <ul>
-    <?php foreach ($errors as $error): ?>
-        <li><?= esc($error) ?></li>
-    <?php endforeach ?>
-    </ul>
-</div>
+<?php
+
+if ($validation->hasError('username')) {
+    echo $validation->getError('username');
+}
