+++ user_guide_src/source/libraries/validation/032.php
@@ -1,7 +1,7 @@
-<?php
-
-public $templates = [
-    'list'    => 'CodeIgniter\Validation\Views\list',
-    'single'  => 'CodeIgniter\Validation\Views\single',
-    'my_list' => '_errors_list',
-];
+<div class="alert alert-danger" role="alert">
+    <ul>
+    <?php foreach ($errors as $error): ?>
+        <li><?= esc($error) ?></li>
+    <?php endforeach ?>
+    </ul>
+</div>
