+++ user_guide_src/source/libraries/validation/036.php
@@ -1,5 +1,9 @@
 <?php
 
-$this->validate($request, [
-    'foo' => 'required|even',
-]);
+class MyRules
+{
+    public function even(string $str): bool
+    {
+        return (int) $str % 2 == 0;
+    }
+}
