+++ user_guide_src/source/libraries/validation/037.php
@@ -1,34 +1,12 @@
 <?php
 
-public function required_with($str, string $fields, array $data): bool
+public function even(string $str, string &$error = null): bool
 {
-    $fields = explode(',', $fields);
+    if ((int) $str % 2 !== 0) {
+        $error = lang('myerrors.evenError');
 
-    // If the field is present we can safely assume that
-    // the field is here, no matter whether the corresponding
-    // search field is present or not.
-    $present = $this->required($str ?? '');
-
-    if ($present) {
-        return true;
+        return false;
     }
 
-    // Still here? Then we fail this test if
-    // any of the fields are present in $data
-    // as $fields is the lis
-    $requiredFields = [];
-
-    foreach ($fields as $field) {
-        if (array_key_exists($field, $data)) {
-            $requiredFields[] = $field;
-        }
-    }
-
-    // Remove any keys with empty values since, that means they
-    // weren't truly there, as far as this is concerned.
-    $requiredFields = array_filter($requiredFields, function ($item) use ($data) {
-        return ! empty($data[$item]);
-    });
-
-    return empty($requiredFields);
+    return true;
 }
