+++ user_guide_src/source/libraries/validation/038.php
@@ -1,10 +1,5 @@
 <?php
 
-// is_unique[table.field,ignore_field,ignore_value]
-
-$validation->setRules([
-    'name' => "is_unique[supplier.name,uuid, $uuid]",  // is not ok
-    'name' => "is_unique[supplier.name,uuid,$uuid ]",  // is not ok
-    'name' => "is_unique[supplier.name,uuid,$uuid]",   // is ok
-    'name' => "is_unique[supplier.name,uuid,{uuid}]",  // is ok - see "Validation Placeholders"
+$this->validate($request, [
+    'foo' => 'required|even',
 ]);
