+++ user_guide_src/source/libraries/validation/040.php
@@ -1,0 +1,10 @@
+<?php
+
+// is_unique[table.field,ignore_field,ignore_value]
+
+$validation->setRules([
+    'name' => "is_unique[supplier.name,uuid, $uuid]",  // is not ok
+    'name' => "is_unique[supplier.name,uuid,$uuid ]",  // is not ok
+    'name' => "is_unique[supplier.name,uuid,$uuid]",   // is ok
+    'name' => "is_unique[supplier.name,uuid,{uuid}]",  // is ok - see "Validation Placeholders"
+]);
